sqlparse,sqlformat: run fidelity corpus through classify and format
Extend fidelity testing from parse-only to classify, format, and
round-trip. TestFidelityClassify asserts every corpus file classifies
without error and produces populated statement fields.
TestFidelityFormat asserts every corpus file formats without error and
round-trips to the same statement count when re-classified.

Closes #56
